changes on onboarding logic
