+++ Fiero/Views/TabBar.swift
@@ -9,16 +9,13 @@ import SwiftUI
 
 struct TabBarView: View {
     
-    let defaults = UserDefaults.standard
-
-
     init() {
         UITabBar.appearance().shadowImage = UIImage()
         UITabBar.appearance().backgroundImage = UIImage()
         UITabBar.appearance().isTranslucent = true
         UITabBar.appearance().backgroundColor = UIColor(Tokens.Colors.Neutral.Low.dark.value)
         UITabBar.appearance().unselectedItemTintColor = UIColor(Tokens.Colors.Neutral.Low.light.value)
-        defaults.set("alreadyOpen", forKey: "isFirstOpen")
+        UserDefaults.standard.set("alreadyOpen", forKey: UDKeys.isFirstOpen.description)
     }
     
     var body: some View {
